+++ components/bl-weather-component/README.md
@@ -11,7 +11,7 @@ Weather is the component that can be used in Backendless [UI-Builder](https://ba
 | Property                                | Type                                                                                                                                | Default value                           | Logic                | Data Binding | UI Setting | Description                                                                                                                                                        |
 |-----------------------------------------|-------------------------------------------------------------------------------------------------------------------------------------|-----------------------------------------|----------------------|--------------|------------|--------------------------------------------------------------------------------------------------------------------------------------------------------------------|
 | Current Location <br> `currentLocation` | *Checkbox*                                                                                                                          | `false`                                 |                      | NO           | YES        | Controls the display of weather by current location.                                                                                                               |
-| City <br> `city`                        | *Text*                                                                                                                              | `New York`                              |                      | NO           | YES        | Controls the city name to display the weather for it.                                                                                                              |
+| City <br> `city`                        | *Text*                                                                                                                              | "New York"                              |                      | NO           | YES        | Controls the city name to display the weather for it.                                                                                                              |
 | Coordinates <br> `coords`               | *JSON*                                                                                                                              | `{"lat": 40.730610, "lng": -73.935242}` | Coordinates Logic    | YES          | YES        | Specifies a JSON object containing the latitude and longitude coordinates data to display the weather for them. Signature of coords: `{lat: Number, lng: Number}`. |
 | Forecast <br> `forecast`                | *Checkbox*                                                                                                                          | `false`                                 |                      | NO           | YES        | Controls the display the weather forecast.                                                                                                                         |
 | Unit <br> `unit`                        | *Select* <br> [`standard`, `metric`, `imperial`]                                                                                    | `standard`                              |                      | NO           | YES        | Controls the unit of measurement.                                                                                                                                  |
@@ -41,6 +41,7 @@ Before usage, you need:
 **Dimensions**
 ```
 @bl-customComponent-weather-size: 100%;
+@bl-customComponent-weather-container-size: 100%;
 ```
 
 ## Theme
