+++ components/bl-weather-component/src/lib/weather.js
