+++ components/bl-weather-component/styles/index.less
@@ -1,9 +1,14 @@
 @bl-customComponent-weather-size: 100%;
+@bl-customComponent-weather-container-size: 100%;
 
 .bl-customComponent-weather {
   width: @bl-customComponent-weather-size;
 
   .rw-container {
-    width: @bl-customComponent-weather-size;
+    width: @bl-customComponent-weather-container-size;
+  }
+
+  .warn {
+    text-align: center;
   }
 }
